telegram_bot/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
gunicorn==22.0.0
idna==3.7
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2024.1
requests==2.32.3
setuptools==70.0.0
six==1.16.0
tornado==6.1
tzlocal==5.2
urllib3==1.26.15
